8c/5c: use %ld and int32 Adr offset, not a long
this helps to display the same value than on -m32
on 8c -S tests/c/misc/vlong_max.c

+++ compilers/5c/gc.h
@@ -31,7 +31,8 @@ typedef	struct	Rgn	Rgn;
 
 struct	Adr
 {
-	long	offset;
+    //old: was long
+	int32	offset;
 	double	dval;
 	char	sval[NSNAME];
 	Ieee	ieee;
